Drop support for KTooltipWidget hack
KTooltipWidget wants to blur the background, we can't use KWindowEffects
in KWidgetAddons a solution using the platform theme and properties was
introduced.

Blurring is very much a property of the style rather than the widget
itself. Especially when it can't be relied upon across desktop.

Given we have code to blur all other windows inside Breeze, it fits far
more naturally there.

We can then drop the property change in KTooltipWidget in KF6.